proposal/pyproject.toml
[project]
name = "proposal-cli"
version = "0.1.0"
description = "Standalone proposal pipeline workspace."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "langgraph>=1.0.7,<2",          
  "langchain-core>=1.2.7,<2",     
  "pydantic>=2.12,<3",           
  "pydantic-ai>=1.47.0,<2",      
  "openai>=2.16.0,<3",           
  "pytest>=7,<9",
  "python-docx",
  "docxtpl",
  "openpyxl",
  "pyyaml",
  "python-dotenv",
  "rank-bm25",
  "jieba",
  "pdfplumber",
  "tomli; python_version < '3.11'",
]


[project.scripts]
proposal-cli = "proposal_cli.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
